restructure application
