+++ app/__init__.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+APP_PATH = Path(__file__).parent.absolute()
