+++ app/static/css/statistics.css
