+++ app/templates/statistics.html
